+++ src/main/java/toyproject/syxxn/back_end/entity/review/ReviewRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReviewRepository extends CrudRepository<Review, Long> {
+public interface ReviewRepository extends CrudRepository<Review, Integer> {
 }
